Canonicalize limiter contracts and Decimal domain types

+++ src/btcbot/adapters/exchange.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
+from decimal import Decimal
 
 from btcbot.domain.accounting import TradeFill
 from btcbot.domain.models import Balance, OpenOrders, Order, OrderSide, OrderSnapshot, PairInfo
@@ -12,7 +13,7 @@ class ExchangeClient(ABC):
         raise NotImplementedError
 
     @abstractmethod
-    def get_orderbook(self, symbol: str, limit: int | None = None) -> tuple[float, float]:
+    def get_orderbook(self, symbol: str, limit: int | None = None) -> tuple[Decimal, Decimal]:
         """Return (best_bid, best_ask)."""
         raise NotImplementedError
 
@@ -37,8 +38,8 @@ class ExchangeClient(ABC):
         self,
         symbol: str,
         side: OrderSide,
-        price: float,
-        quantity: float,
+        price: Decimal,
+        quantity: Decimal,
         client_order_id: str | None = None,
     ) -> Order:
         raise NotImplementedError
